+++ internal/aghos/os.go
@@ -4,10 +4,16 @@
 package aghos
 
 import (
+	"bufio"
 	"fmt"
+	"io"
 	"os/exec"
+	"path"
 	"runtime"
-	"syscall"
+	"strconv"
+	"strings"
+
+	"github.com/AdguardTeam/golibs/log"
 )
 
 // UnsupportedError is returned by functions and methods when a particular
@@ -43,11 +49,6 @@ func HaveAdminRights() (bool, error) {
 	return haveAdminRights()
 }
 
-// SendProcessSignal sends signal to a process.
-func SendProcessSignal(pid int, sig syscall.Signal) error {
-	return sendProcessSignal(pid, sig)
-}
-
 // MaxCmdOutputSize is the maximum length of performed shell command output.
 const MaxCmdOutputSize = 2 * 1024
 
@@ -63,6 +64,91 @@ func RunCommand(command string, arguments ...string) (int, string, error) {
 	}
 
 	return cmd.ProcessState.ExitCode(), string(out), nil
+}
+
+// PIDByCommand searches for process named command and returns its PID ignoring
+// the PIDs from except.  If no processes found, the error returned.
+func PIDByCommand(command string, except ...int) (pid int, err error) {
+	// Don't use -C flag here since it's a feature of linux's ps
+	// implementation.  Use POSIX-compatible flags instead.
+	//
+	// See https://github.com/AdguardTeam/AdGuardHome/issues/3457.
+	cmd := exec.Command("ps", "-A", "-o", "pid=", "-o", "comm=")
+
+	var stdout io.ReadCloser
+	if stdout, err = cmd.StdoutPipe(); err != nil {
+		return 0, fmt.Errorf("getting the command's stdout pipe: %w", err)
+	}
+
+	if err = cmd.Start(); err != nil {
+		return 0, fmt.Errorf("start command executing: %w", err)
+	}
+
+	var instNum int
+	pid, instNum, err = parsePSOutput(stdout, command, except...)
+	if err != nil {
+		return 0, err
+	}
+
+	if err = cmd.Wait(); err != nil {
+		return 0, fmt.Errorf("executing the command: %w", err)
+	}
+
+	switch instNum {
+	case 0:
+		// TODO(e.burkov):  Use constant error.
+		return 0, fmt.Errorf("no %s instances found", command)
+	case 1:
+		// Go on.
+	default:
+		log.Info("warning: %d %s instances found", instNum, command)
+	}
+
+	if code := cmd.ProcessState.ExitCode(); code != 0 {
+		return 0, fmt.Errorf("ps finished with code %d", code)
+	}
+
+	return pid, nil
+}
+
+// parsePSOutput scans the output of ps searching the largest PID of the process
+// associated with cmdName ignoring PIDs from ignore.  Valid r's line shoud be
+// like:
+//
+//    123 ./example-cmd
+//   1230 some/base/path/example-cmd
+//   3210 example-cmd
+//
+func parsePSOutput(r io.Reader, cmdName string, ignore ...int) (largest, instNum int, err error) {
+	s := bufio.NewScanner(r)
+ScanLoop:
+	for s.Scan() {
+		fields := strings.Fields(s.Text())
+		if len(fields) != 2 || path.Base(fields[1]) != cmdName {
+			continue
+		}
+
+		cur, aerr := strconv.Atoi(fields[0])
+		if aerr != nil || cur < 0 {
+			continue
+		}
+
+		for _, pid := range ignore {
+			if cur == pid {
+				continue ScanLoop
+			}
+		}
+
+		instNum++
+		if cur > largest {
+			largest = cur
+		}
+	}
+	if err = s.Err(); err != nil {
+		return 0, 0, fmt.Errorf("scanning stdout: %w", err)
+	}
+
+	return largest, instNum, nil
 }
 
 // IsOpenWrt returns true if host OS is OpenWrt.