+++ internal/aghos/os_freebsd.go
@@ -20,10 +20,6 @@ func haveAdminRights() (bool, error) {
 	return os.Getuid() == 0, nil
 }
 
-func sendProcessSignal(pid int, sig syscall.Signal) error {
-	return syscall.Kill(pid, sig)
-}
-
 func isOpenWrt() (ok bool) {
 	return false
 }