+++ CHANGELOG.md
@@ -65,6 +65,7 @@ and this project adheres to
 
 ### Fixed
 
+- `reload` service action on macOS and FreeBSD ([#3457]).
 - Inaccurate using of service actions in the installation script ([#3450]).
 - Client ID checking ([#3437]).
 - Discovering other DHCP servers on `darwin` and `freebsd` ([#3417]).
@@ -128,6 +129,7 @@ and this project adheres to
 [#3435]: https://github.com/AdguardTeam/AdGuardHome/issues/3435
 [#3437]: https://github.com/AdguardTeam/AdGuardHome/issues/3437
 [#3450]: https://github.com/AdguardTeam/AdGuardHome/issues/3450
+[#3457]: https://github.com/AdguardTeam/AdGuardHome/issues/3457